feat: add login migrations features
