+++ .env
@@ -1,2 +1,3 @@
 DB_URL="postgres://postgres:1234@localhost:5432/desafio"
 API_PORT=":3000"
+TOKEN_SECRET="bags lime ruts loop dale muse"
