+++ pkg/config/config.go
@@ -1,0 +1,7 @@
+package config
+
+type Config struct {
+	DatabaseURL string `envconfig:"DB_URL" required:"true"`
+	Port        string `envconfig:"API_PORT" required:"true" default:":3000"`
+	TokenSecret string `envconfig:"TOKEN_SECRET" required:"true"`
+}
