+++ pkg/domain/login/use_case.go
@@ -2,8 +2,6 @@ package login
 
 import (
 	"context"
-
-	"github.com/daniel1sender/Desafio-API/pkg/domain/entities"
 )
 
 type LoginUseCase struct {
@@ -13,10 +11,6 @@ type LoginUseCase struct {
 
 type UseCase interface {
 	Auth(ctx context.Context, cpf, secret string) (string, error)
-}
-
-type AccountRepository interface {
-	GetByCPF(ctx context.Context, cpf string) (entities.Account, error)
 }
 
 func NewUseCase(accountStorage AccountRepository, tokenSecret string) LoginUseCase {
