+++ run.sh
@@ -1,4 +1,5 @@
 #%/bin/bash
 export DB_URL="postgres://postgres:1234@localhost:5432/desafio"
 export API_PORT=":3000"
+export TOKEN_SECRET="bags lime ruts loop dale muse"
 go run main.go
